+++ app/src/main/res/values/strings.xml
@@ -17,4 +17,5 @@
     <string name="hours_colon">Hours:</string>
     <string name="overtime_colon">Overtime:</string>
     <string name="penalty_colon">Penalty:</string>
+    <string name="empty_not_saved">Day not saved because it is empty.</string>
 </resources>
